domainError entity file renamed to DomainError decorator domainError status code changed to 400 added a real try-catch example
