+++ src/domain/entities/DomainError.ts
